requirements/base.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.41.0
django-braces==1.14.0
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-embed-video==1.3
django-environ==0.4.5
django-filter==2.2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-login-required-middleware==0.4
django-model-utils==3.2.0
django-mptt==0.10.0
django-widget-tweaks==1.4.5
idna==2.8
mpmath==1.1.0
oauthlib==3.1.0
olefile==0.46
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
sympy==1.5.1
transliterate==1.10.2
urllib3==1.25.7
